==> Bark_Beetle_UAS/src/03_Feature_Extraction.R <==
######################################################
# Bark_Beetle_UAS: Detection of bark beetle infestation in
# spruce plantations using multispectral drone images
#
# 03 Feature Extraction
# Forest Inventory and Remote Sensing
# University Göttingen
# Dr. Hans Fuchs
# hfuchs@gwdg.de
#######################################################

# set working directory
# adapt to your local directory where the project is stored.
wd = "~/Bark_Beetle_UAS"
setwd(wd)

#######################################################
# References
# Freudenberg, M., Magdon, P., Nölke, N. (2022): 
# Individual tree crown delineation in high-resolution remote sensing
# images based on U-Net. Neural Computing and Applications. 
# https://doi.org/10.1007/s00521-022-07640-4
#######################################################


#######################################################
# Importing libraries
#######################################################

library(raster)
library(rgdal)
library(sf)
library(dplyr)
library(exactextractr)


# import segments generated by Max Freudenberg (s. Freudenberg et al. (2022))  
tpoly = st_read(dsn= "./data/clip_RedEdge_Mission_polygons.gpkg")

# import prepropcessed uas image stack
uas_vi = raster::brick('./data/UAS_VI.tif')
names(uas_vi) = c('blue','green','red','re','nir','rendvi','gndvi','ratio','cgm','crem')
NAvalue(uas_vi) = 65535
uas_vi

# extract features of pixel values within each tree polygon returning a data frame
tpoly.train = exact_extract(uas_vi,tpoly,c('mean','stdev'),append_cols='id',progress=TRUE, max_cells_in_memory =123758401)

tpolygons = merge(tpoly, tpoly.train, by = "id")

#Save the training data 
st_write(tpolygons, dsn = "./data/tpolygons.gpkg", driver = "GPKG")

# clear up memory
rm(list = ls())
gc()
